Fix import of error controller and improve timer code
Previously when using controller packages, if an exception was
raised, the error controller was imported from the package location of
the controller itself.  The solution was to set the package lcoation
for the error controller to be the default classpath.  The timer code
was updated to no longer append the timer after the </body> tag, now
it's xhtml compliant - and potentially a little faster.