Fix corner case in FieldStorage array structures
Usually when either mod_python or cgi's FieldStorage classes encounter
an array type it's accessed via a "value" attribute.  In these cases
the object type is Field or MiniFieldStorage, but sometimes it's a
string for some reason.  I've not been able to reproduce it - but I've
seen it in my logs - so I know it happens.