Initial revisit of the queue
Updated the mqueue class to actually inherit from Queue.Queue.  This
allowed me to stop polling for new messages, and always creating a new
thread per message.  Now I create a fixed number of worker threads
(the number is calculated from the number of cores you have) which
then consume messages in realtime.  This is far more efficent and
easier to mange.

One thing that's currently missing is the ability to recover from
failure.  Previously the poller would just pick up a message that got
skipped.  Now a new thread needs to be fired at startup that just
walks the directory for any messages and add them to the queue (before
the socket gets opened up).