Fix session regression cause by recent memory work
In the session._gc() method I was setting the memcached cache instance
to None so it would be unbound.  Unfortunately this was getting called
when earlier than I realized (on db persist) which meant that later,
when fetching from the cache it couldn't... and did a force persist.

Fixed these problems, and added logging for use with working on this
stuff as it's a bit tricky.  Also made the number of reqeusts before a
commit to the db configurable (session_max_stale_count).