Add support for custom queue messages to be used
For a while now the class type was represented by the name of the
class.  There was then a hard coded check against this to see if it
was a valid class.  Now the entire classpath is persisted, and a check
is made to ensure it's a subclass of Message().  This makes the code
much more maintainable, as well as making it super easy for anyone to
push their custom messages thru the queue.  It's also nice to just let
Python handle the object types rather than trying to type check things
ourselves.  Currently the only type related stuff we're handling is
persisting the name inside the encoded structure.  I presume pickling
would essentially do the same thing, and I'm hoping to keep this
Python agnostic so any tcp enabled clent can generate messages.

Closes GH-7