Initial import of regex based mapper
Also updated wsgi env to include querystring as part of the
REQUEST_URI when not specified upstream.  This is consistent with the
spirit of what REQUEST_URI is supposed to hold, per the WSGI spec.

Currently trying to sort out the best way to unit/bat test the two
mapers that are now supported.