Updates after an ubuntu vm dry run
This sets the required python version to be 2.6, and roadmap comments
about 2.5.  Updated the logger to not use the console logger unless
debugging is turned on (oops).  Moved the session backend imports such
that they only get imported when used - this means you only need the
drivers for the backends in use.  Fixed a missing comma in the nginx
config fragment, as well as setting perms on the fastcgi socket.
Removed the commented singleton on the session class, as this isn't a
singleton at all.