Fix post_raw regression from clean/downcast returns
Initially there was an issue where the code worked when given args,
but broke with raw input when it tried to process the post args.  This
occured because the code expected to deal with a dict, and json/xml is
not that.  The fix was to just return early (aka noop) the _clean and
_downcast methods (called by extras()).

Unfortunately this caused another issue as the logic wasn't tight
enough.  Because the downcast method was basically skipped, the
form object didn't get converted to a dict() and lacked the .get()
method.  This broke inside chula.webservice.expose().

The more reliable fix is to more carefully consider the data before
building the structures.  Now the HTTP GET data is processed first,
regardless.  Then the form data is processed (minus the keys in the
GET data).  If there is only one key, and it smells like xml or json,
then it's not really a key... it's a raw string post.  In this case we
reset the structures accordingly.

This still hasn't been tested extensively and unit tests need to be
created prior to this merging in.