Rename queue exception to be more accurate
When a successfully processed message is purged from the queue, the
file on disk gets moved into the appropriate directory.  In the (odd)
event that the message was processed (even if it raised an exception)
but the purging is unable to successfully purge due to an IOError, the
specific error type raised was poorly named.  The exception has been
renamed to be more accurate/clear/specific:

before: CannotPurgeUnprocessedError
after:  IOErrorWhenPurgingProcessedMessage