Set psycopg2.extensions.{UNICODE,UNICODEARRAY}
By default psycopg2 tries to return strings encoded in whatever the
database encoding is set to.  This means the resturn objects are of
type <str>.  When these strings are passed along, it's unknown what
encoding they were encoded with - resulting in crazy.  By always
ensuring data returned by psygopg2 is of type unicode, the code
downstream can just encode it however it needs to.  This makes life
much easier.