Cleanup doc/test execution to handle server startup
For a while now the unit tests would start up a server to test
against, but it often conflicted with the default port.  Additionally
the documentation generation script (which tests the documentation)
also depends on a server being up, but didn't start one automatically.
Everything is now fully automatic, with the "test tcp port" defined in
a single location for easy maintenance.

Also deleted the webserver.py script (which the test case launcher
used to use to startup the server) as it's been deprecated in favor of
chula-run for a while now.